// frontend/src/app/api/model/whLocationDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface WhLocationDto {
  areaId?: number;
  name?: string;
  level?: number;
  children?: Array<WhLocationDto>;
}